# src/main/java/pro/sky/adsplatform/dto/FullAdsDto.java
package pro.sky.adsplatform.dto;
import lombok.Data;

@Data
public class FullAdsDto {
  private String authorFirstName;
  private String authorLastName;
  private String description;
  private String email;
  private String image;
  private String phone;
  private Integer pk;
  private Integer price;
  private String title;
}